refactor: merge WorkspaceController into AdminController and update related methods and views
