+++ src/main/resources/WEB-INF/views/admin.html
@@ -45,7 +45,7 @@
             <td th:text="${workspace.price}"></td>
             <td th:text="${workspace.address}"></td>
             <td>
-                <form action="#" method="post" th:action="@{'/workspace/delete/' + ${workspace.id}}">
+                <form action="#" method="post" th:action="@{'/admin/delete-workspace/' + ${workspace.id}}">
                     <button class="btn btn-outline-warning" data-bs-toggle="modal"
                             th:data-bs-target="@{'#editWorkspaceModal' + ${workspace.id}}" type="button"><i
                             class="bi bi-pen"></i>
@@ -65,7 +65,7 @@
                                 <button aria-label="Close" class="btn-close" data-bs-dismiss="modal"
                                         type="button"></button>
                             </div>
-                            <form action="#" method="post" th:action="@{'/workspace/edit/' + ${workspace.id}}"
+                            <form action="#" method="post" th:action="@{'/admin/edit-workspace/' + ${workspace.id}}"
                                   th:object="${editWorkspace}">
                                 <div class="modal-body">
                                     <div class="row">
@@ -161,7 +161,7 @@
                 <h5 class="modal-title" id="createWorkspaceTitle">Create Workspace</h5>
                 <button aria-label="Close" class="btn-close" data-bs-dismiss="modal" type="button"></button>
             </div>
-            <form action="#" method="post" th:action="@{/workspace/create}" th:object="${createWorkspace}">
+            <form action="#" method="post" th:action="@{/admin/create-workspace}" th:object="${createWorkspace}">
                 <div class="modal-body">
                     <div class="row">
                         <div class="col">
